split useRoomList and useSubscriptions
