+++ apps/meteor/client/sidebar/RoomList/RoomList.tsx
@@ -1,4 +1,4 @@
-import type { IRoom } from '@rocket.chat/core-typings';
+import type { IRoom, ISubscription } from '@rocket.chat/core-typings';
 import { css } from '@rocket.chat/css-in-js';
 import { Box } from '@rocket.chat/fuselage';
 import { useResizeObserver } from '@rocket.chat/fuselage-hooks';
@@ -7,21 +7,36 @@ import type { ReactElement } from 'react';
 import React, { useMemo } from 'react';
 import { Virtuoso } from 'react-virtuoso';
 
+import { CachedChatRoom } from '../../../app/models/client/models/CachedChatRoom';
+import { CachedChatSubscription } from '../../../app/models/client/models/CachedChatSubscription';
 import { useOpenedRoom } from '../../lib/RoomManager';
 import { useAvatarTemplate } from '../hooks/useAvatarTemplate';
 import { usePreventDefault } from '../hooks/usePreventDefault';
 import { useRoomList } from '../hooks/useRoomList';
 import { useShortcutOpenMenu } from '../hooks/useShortcutOpenMenu';
+import { useSubscriptions } from '../hooks/useSubscriptions';
 import { useTemplateByViewMode } from '../hooks/useTemplateByViewMode';
 import RoomListRow from './RoomListRow';
 import ScrollerWithCustomProps from './ScrollerWithCustomProps';
 
 const computeItemKey = (index: number, room: IRoom): IRoom['_id'] | number => room._id || index;
 
+const cachedCollectionFunctions = {
+	handleSubscription: (action: 'inserted' | 'updated' | 'removed', record: ISubscription) =>
+		CachedChatSubscription.handleEvent(action === 'removed' ? action : 'changed', record),
+	handleRoom: (action: 'inserted' | 'updated' | 'removed', record: IRoom) =>
+		CachedChatRoom.handleEvent(action === 'removed' ? action : 'changed', record),
+	applyFromServer: (subscriptions: ISubscription[], rooms: IRoom[]) => {
+		CachedChatSubscription.applyFromServer(subscriptions);
+		CachedChatRoom.applyFromServer(rooms);
+	},
+};
+
 const RoomList = (): ReactElement => {
 	const t = useTranslation();
 	const isAnonymous = !useUserId();
-	const roomsList = useRoomList();
+	const userSubscriptions = useSubscriptions(cachedCollectionFunctions);
+	const roomsList = useRoomList(userSubscriptions.data ?? []);
 	const avatarTemplate = useAvatarTemplate();
 	const sideBarItemTemplate = useTemplateByViewMode();
 	const { ref } = useResizeObserver({ debounceDelay: 100 });
@@ -120,7 +135,7 @@ const RoomList = (): ReactElement => {
 			<Box h='full' w='full' ref={ref}>
 				<Virtuoso
 					totalCount={roomsList.length}
-					data={roomsList}
+					data={roomsList as unknown as (ISubscription & IRoom)[]}
 					components={{ Scroller: ScrollerWithCustomProps }}
 					computeItemKey={computeItemKey}
 					itemContent={(_, data): ReactElement => <RoomListRow data={itemData} item={data} />}
