+++ ee/packages/ddp-client/src/types/streams.ts
@@ -69,7 +69,7 @@ export interface StreamerEvents {
 	];
 
 	'notify-user': [
-		{ key: `${string}/rooms-changed`; args: ['inserted' | 'updated' | 'removed' | 'changed', IRoom] },
+		{ key: `${string}/rooms-changed`; args: ['inserted' | 'updated' | 'removed', IRoom] },
 		{
 			key: `${string}/subscriptions-changed`;
 			args:
